2011 posts with theme updates for proper code blocks
Add just the first year of posts and make sure they look okay.

This required two changes to the theme to get code blocks to look okay:

* don't wrap lines of code, since this throws the line numbers off and makes
everything look horrible
* get rid of bottom margin for `pre` elements, which caused a weird white space
to appear between the border at the bottom of a code block and the scrollbar.

I sent PR's for both of these changes (
https://github.com/mpas/hugo-multi-bootswatch/pull/10 and
https://github.com/mpas/hugo-multi-bootswatch/pull/11), so hopefully they'll be
merged and I won't have to maintain a fork.